refactor(Orders): long을 int로 변경

+++ src/main/java/christmas/domain/Orders.java
@@ -24,8 +24,8 @@ public class Orders {
                 .sum();
     }
 
-    public long getCourseCount(Course course) {
-        return orders.stream()
+    public int getCourseCount(Course course) {
+        return (int) orders.stream()
                 .filter(order -> order.is(course))
                 .count();
     }
